ensure directory.

+++ scripts/rmlinks.sh
@@ -1,7 +1,9 @@
 #!/bin/bash
-cd "$1"
-for i in pie-chart logs compilers config; do
-    rm -f "$i"
-done
-rm -f ./htaccess
-rm -f ./favicon.png
+if [ -d "$1" ]; then
+    cd "$1"
+    for i in pie-chart logs compilers config; do
+        rm -f "$i"
+    done
+    rm -f ./htaccess
+    rm -f ./favicon.png
+fi
